+++ src/core_modules/capture-core/components/Pages/common/TEIRelationshipsWidget/RegisterTei/RegisterTei.types.js
@@ -1,0 +1,23 @@
+// @flow
+export type SharedProps = {|
+    onLink: (teiId: string, values: Object) => void,
+    onGetUnsavedAttributeValues?: ?Function,
+    trackedEntityTypeId: string,
+|};
+
+export type ContainerProps = {|
+    suggestedProgramId: string,
+    onSave: (teiPayload: Object) => void,
+    ...SharedProps,
+|};
+
+export type ComponentProps = {|
+    selectedScopeId: string,
+    error: string,
+    dataEntryId: string,
+    trackedEntityName: ?string,
+    onSaveWithEnrollment: () => void,
+    onSaveWithoutEnrollment: () => void,
+    ...SharedProps,
+    ...CssClasses,
+|};