+++ src/core_modules/capture-core/components/WidgetsRelationship/WidgetTrackedEntityRelationship/NewTrackedEntityRelationship/common/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { TARGET_SIDES } from './targetSides';