+++ src/core_modules/capture-core/components/WidgetsRelationship/WidgetTrackedEntityRelationship/NewTrackedEntityRelationship/common/targetSides.js
@@ -1,0 +1,9 @@
+// @flow
+
+export const TARGET_SIDES: {|
+    FROM: 'FROM',
+    TO: 'TO',
+|} = Object.freeze({
+    FROM: 'FROM',
+    TO: 'TO',
+});