+++ src/core_modules/capture-core/reducers/descriptions/organisationUnits.reducerDescription.js
@@ -24,5 +24,5 @@ export const organisationUnitRootsDesc = createReducerDescription({
             roots: action.payload.roots,
         },
     }),
-}, 'organisationUnitRoots');
+}, 'organisationUnitRoots2222');